requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
lxml==4.9.0
pip==22.1.2
python-dateutil==2.8.2
requests==2.28.0
setuptools==62.6.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
wheel==0.37.1
airium~=0.2.3
xhtml2pdf~=0.2.8